Add functions to merge Valgo errors from validated sessions
This commit introduces three new functions for merging Valgo errors
produced by other validation sessions. These functions make it possible
to merge errors in a field key, as well as in a namespace or indexed
namespace. The added functions are:

* MergeError: Merges Valgo errors from an already validated Validation session.
* MergeErrorIn: Merges Valgo errors from already validated Validation sessions
within an indexed namespace.
* MergeErrorInRow: Merges Valgo errors from already validated Validation
sessions within a map namespace. These enhancements improve the flexibility of
error handling and management within the Valgo library.